--- src/main/java/webprogrammingTeam/matchingService/domain/Review/dto/response/ReviewAllReadResponse.java
@@ -1,8 +1,0 @@
-package webprogrammingTeam.matchingService.domain.Review.dto.response;
-
-import lombok.Builder;
-import webprogrammingTeam.matchingService.domain.Review.entity.Rating;
-
-@Builder
-public record ReviewAllReadResponse(Rating rating, String content, String date) {
-}
