--- src/main/java/webprogrammingTeam/matchingService/domain/Review/respository/ReviewRepository.java
@@ -1,14 +1,0 @@
-package webprogrammingTeam.matchingService.domain.Review.respository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import webprogrammingTeam.matchingService.domain.Review.entity.Review;
-import webprogrammingTeam.matchingService.domain.board.entity.Board;
-import webprogrammingTeam.matchingService.domain.user.entity.User;
-
-import java.util.List;
-import java.util.Optional;
-
-public interface ReviewRepository extends JpaRepository<Review, Long> {
-
-    List<Review> findByBoard(Board board);
-}
