+++ src/main/java/webprogrammingTeam/matchingService/domain/member/entity/Role.java
@@ -1,0 +1,8 @@
+package webprogrammingTeam.matchingService.domain.member.entity;
+
+import lombok.Getter;
+
+@Getter
+public enum Role {
+    ROLE_GUEST, ROLE_MEMBER, ROLE_ADMIN
+}
