+++ src/main/java/webprogrammingTeam/matchingService/domain/refresh/entity/RefreshEntity.java
@@ -16,7 +16,7 @@ public class RefreshEntity {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private String username;
+    private String memberName;
     private String refresh;
     private String expiration;
