+++ src/main/java/webprogrammingTeam/matchingService/domain/review/dto/request/ReviewSaveRequest.java
@@ -1,0 +1,8 @@
+package webprogrammingTeam.matchingService.domain.review.dto.request;
+
+import lombok.Builder;
+import webprogrammingTeam.matchingService.domain.review.entity.Rating;
+
+@Builder
+public record ReviewSaveRequest(Rating rating, String content, String date){
+}
