+++ src/main/java/webprogrammingTeam/matchingService/domain/review/dto/request/ReviewUpdateRequest.java
@@ -1,0 +1,8 @@
+package webprogrammingTeam.matchingService.domain.review.dto.request;
+
+import lombok.Builder;
+import webprogrammingTeam.matchingService.domain.review.entity.Rating;
+
+@Builder
+public record ReviewUpdateRequest(Rating rating, String content) {
+}
