+++ src/main/java/webprogrammingTeam/matchingService/domain/review/dto/response/ReviewAllReadResponse.java
@@ -1,0 +1,8 @@
+package webprogrammingTeam.matchingService.domain.review.dto.response;
+
+import lombok.Builder;
+import webprogrammingTeam.matchingService.domain.review.entity.Rating;
+
+@Builder
+public record ReviewAllReadResponse(Rating rating, String content, String date) {
+}
