--- src/main/java/webprogrammingTeam/matchingService/domain/subscription/dto/UserChannelSubscriptionDTO.java
@@ -1,13 +1,0 @@
-package webprogrammingTeam.matchingService.domain.subscription.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class UserChannelSubscriptionDTO {
-
-    private Long userId;
-
-    private Long channelId;
-}
