--- src/main/java/webprogrammingTeam/matchingService/domain/subscription/entity/UserChannelSubscription.java
@@ -1,27 +1,0 @@
-package webprogrammingTeam.matchingService.domain.subscription.entity;
-
-import webprogrammingTeam.matchingService.domain.user.entity.User;
-import webprogrammingTeam.matchingService.domain.channel.entity.Channel;
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.Setter;
-
-@Entity
-@Getter
-@Setter
-public class UserChannelSubscription {
-    // 유저와 채팅방 사이의 다대다 관계
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long subscriptionId;
-
-    @ManyToOne
-    @JoinColumn(name = "user")
-    private User user;
-
-    @ManyToOne
-    @JoinColumn(name = "channel")
-    private Channel channel;
-
-}
