--- src/main/java/webprogrammingTeam/matchingService/domain/user/entity/Role.java
@@ -1,8 +1,0 @@
-package webprogrammingTeam.matchingService.domain.user.entity;
-
-import lombok.Getter;
-
-@Getter
-public enum Role {
-    ROLE_GUEST, ROLE_USER, ROLE_ADMIN
-}
