--- src/main/java/webprogrammingTeam/matchingService/domain/user/repository/UserRepository.java
@@ -1,10 +1,0 @@
-package webprogrammingTeam.matchingService.domain.user.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import webprogrammingTeam.matchingService.domain.user.entity.User;
-
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User,Long> {
-    Optional<User> findByEmail(String email);
-}
