+++ src/main/resources/application.yml
@@ -1,0 +1,38 @@
+server:
+  port: 8080
+spring:
+  security:
+    oauth2:
+      client:
+        registration:
+          google:
+            client-id: 260106878823-obralj8i3rel9nmr8ugbdvj6v6gqkmge.apps.googleusercontent.com
+            client-secret: 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
+            redirect-uri: http://localhost:8080/login/oauth2/code/google
+            scope: profile, email
+
+  jpa:
+    hibernate:
+      ddl-auto: create
+    properties:
+      hibernate:
+        show_sql: true
+        format_sql: true
+        dialect: org.hibernate.dialect.MySQLDialect
+  datasource:
+    driver-class-name: com.mysql.cj.jdbc.Driver
+    url: jdbc:mysql://localhost:3306/MYSQL?useSSL=false&serverTimezone=Asia/Seoul&characterEncoding=UTF-8
+    username: root
+    password: 16tngh%%
+  jwt:
+    secret: kjvMLWXxW4VDgWK6Cf5QjFiDoCRrQfKDFZQQgmT8gzF2pwfQPy
+
+logging:
+  level:
+    org:
+    hibernate:
+      SQL: DEBUG
+    type:
+      descriptor:
+        sql:
+          BasicBinder: TRACE
